requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
reportlab==4.0.4
PyPDF2==3.0.1
Flask-Mail==0.9.1
requests==2.31.0
SQLAlchemy==2.0.23
alembic==1.12.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
python-dateutil==2.8.2
pytz==2023.3
pytest
pytest-flask
Flask-RESTful
PyJWT 